+++ compiler/CHANGELOG.md
@@ -7,6 +7,8 @@ This changelog summarizes newly introduced optimizations that may be relevant to
 that iterate in 32bit ranges.
 * (GR-29341) AVX-512 support: Fix EVEX encoding and feature checks for existing instructions and add AVX-512 
 alternatives.
+* (GR-31162) Do not de-duplicate ValueAnchorNode. As part of this change, there is a new marker interface
+NodeWithIdentity to mark nodes that have identity.
 
 ## Version 21.1.0
 * (GR-29126) Unify box optimizations in the compiler. Remove `-Dgraal.ReuseOutOfCacheBoxedValues=false`.